# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiKeyMetadata": {
        "description": "Safe API-key metadata \u2014 never the ``key_hash`` or the plaintext secret (Req 5.4).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "key_prefix": {
            "title": "Key Prefix",
            "type": "string"
          },
          "org_id": {
            "format": "uuid",
            "title": "Org Id",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "org_id",
          "role",
          "key_prefix",
          "created_at"
        ],
        "title": "ApiKeyMetadata",
        "type": "object"
      },
...
      "CreateApiKeyRequest": {
        "description": "Body for ``POST /orgs/{id}/api-keys`` \u2014 issue a key granting ``role`` (Req 5.1).",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "role"
        ],
        "title": "CreateApiKeyRequest",
        "type": "object"
      },
...
      "CreateOrgRequest": {
        "description": "Body for ``POST /orgs`` \u2014 an authenticated user creates an org they own.",
        "properties": {
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateOrgRequest",
        "type": "object"
      },
      "CreateOrgResponse": {
        "description": "Response carrying the newly-created Org_Id (Req 2.1).",
        "properties": {
          "org_id": {
            "format": "uuid",
            "title": "Org Id",
            "type": "string"
          }
        },
        "required": [
          "org_id"
        ],
        "title": "CreateOrgResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AgentForge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/orgs": {
      "post": {
        "description": "Create a NEW Organization owned by the caller (Req 2.1, 2.2).\n\nAny authenticated principal may create an org; a User principal is added as its\n``OWNER`` so they can immediately manage it.",
        "operationId": "create_org_orgs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrgResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Org",
        "tags": [
          "orgs"
        ]
      }
    },
    "/orgs/{org_id}/api-keys": {
      "get": {
        "description": "Return metadata for every API key in ``org_id`` \u2014 never a hash or secret (Req 5.4).",
        "operationId": "list_api_keys_orgs__org_id__api_keys_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyMetadata"
                  },
                  "title": "Response List Api Keys Orgs  Org Id  Api Keys Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Api Keys",
        "tags": [
          "orgs"
        ]
      },
      "post": {
        "description": "Issue an org-scoped API key; return the plaintext secret **once** (Req 5.1, 5.2).",
        "operationId": "create_api_key_orgs__org_id__api_keys_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Api Key",
        "tags": [
          "orgs"
        ]
      }
    },
    "/orgs/{org_id}/api-keys/{key_id}": {
      "delete": {
        "description": "Revoke an API key; an unknown or cross-org key is a 404 (Req 5.5, 5.7).",
        "operationId": "revoke_api_key_orgs__org_id__api_keys__key_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Api Key",
        "tags": [
          "orgs"
        ]
      }
    },
    "/orgs/{org_id}/members": {
      "get": {
        "description": "Return every Membership in ``org_id`` with its User's email (Req 2.6, 4.4).\n\nThe store filters by ``org_id``, so another tenant's roster is structurally\nunreachable; emails are resolved in one batched call rather than per row.",
        "operationId": "list_members_orgs__org_id__members_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberSummary"
                  },
                  "title": "Response List Members Orgs  Org Id  Members Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Members",
        "tags": [
          "orgs"
        ]
      },
      "post": {
        "description": "Add an existing user (by email) to ``org_id`` under a Role (Req 2.2).",
        "operationId": "add_member_orgs__org_id__members_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddMemberResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Member",
        "tags": [
          "orgs"
        ]
      }
    },
    "/orgs/{org_id}/members/{user_id}": {
      "delete": {
        "description": "Remove a member from ``org_id``, together with their Team_Memberships (Req 2.5).\n\nAn unknown or cross-tenant member is a uniform 404; removing the last owner raises\n``AppError(\"last_owner\", 400)``.",
        "operationId": "remove_member_orgs__org_id__members__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Member",
        "tags": [
          "orgs"
        ]
      },
      "patch": {
        "description": "Reassign a member's Role within ``org_id`` (Req 2.2, 3.4).\n\nAn unknown member \u2014 or one belonging to another tenant \u2014 is a uniform 404. Demoting\nthe organization's last owner raises ``AppError(\"last_owner\", 400)``.",
        "operationId": "update_member_role_orgs__org_id__members__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Member Role",
        "tags": [
          "orgs"
        ]
      }
    },
...
    "/orgs/{org_id}/teams/{team_id}": {
      "delete": {
        "description": "Delete a Team from ``org_id`` with its Team_Memberships; unknown/cross-org is 404.",
        "operationId": "delete_team_orgs__org_id__teams__team_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Team Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Team",
        "tags": [
          "orgs"
        ]
      }
    },
...
  }
}